Database/Migrations: move register commands from ArtisanServiceProvider -> MigrationServiceProvider
